Move pkg/subctl/operator/common/deployments to pkg/deployment
Also rename WaitForReady to AwaitReady.
